Moved slice decoding into different file
